Added a simple general purpose LRU cache in src/cache.{h,c} and used it for a FileRecord cache (http://mongrel2.com/tktview?name=2aaa92fa9e).
Changed the semantics of FileRecord so that now you find/release a file (rather
than find/destroy).  Destroy is called when the number of users (including the
cache) reaches 0.  This is good, because it means that we don't have to worry
about closing the file during a long-lasting operation because we've evicted
it from the cache.

Finally, added a testfile (tests/cache_tests.c) to verify that the LRU cache 
works as expected.  I also wrote a really simple test using javascript to fetch
ranges of files (woo jquery), and used a debug statement for hits/misses to
make sure that it all operated as expected.  I'll add a bigger test to the dir
tests later.