Change websockets to pass opening handshake on
This allows the client to negotiate extensions.  We don't fully support
extensions yet, since mongrel2 always defragments packets, which is
disallowed in the presence of unknown extensions.  On the other hand,
defragmenting packets in e.g. python is incredibly slow vs. C so I think
it's a resonable middle ground right now.

Note that if someone sends an opening handshake, followed by a websocket
frame, we will start processing those frames regardless of whether or
not the handler signals it supports websockets by replying with a
response.  This *may* be a security vulnerability, I haven't determined
that yet.