Remove set/get/make/swapcontext implementations for arm linux
These implemtations are not needed any longer, as arm linux does provide
them in glibc 2.17. Implementations for uClibc are available, as well.

This also fixes zedshaw/mongrel2#232, where mongrel2 on armhf was broken
because these implementations don't support that architecture, but where
enabled because the code only checks for __arm__.