Ensure the full set of tokens are copied into a Command
The maximum number of tokens (as opposed to the number of bytes) was used as
the length argument to a memcpy() call.

In practice this didn't turn out to be an issue since it would require more
than 13 or 25 (MAX_TOKENS / sizeof(Token *)) command line arguments to trigger,
and even then it would result in observably correct behavior since the
memcpy()'s destination was only used in Command_destroy() which would simply
leak the un-copied tokens.