Start to automate release process
This includes a lot of infrastructure change.

I also added a rockspec for just the templating part of Tir, which is
all that is needed to build the site.  Hopefully people that aren't me
will be able to build the site now.

Also fixed one place in the manual that still referenced the old github
location.