Fix bug in SuperPoll when socket errors happen
There was an underlying assumption that the return value of zmq_poll
would be equal to the number of items with revents set.  This is not
actually true, as polling for read/write without error can return when
the sockets are in an error state, with no bits set in revents.

This causes it to read past the end of the valid entries in the data
array.

It is suspected to be a cause of the crash in #316