Fix many bugs in signal handler
The MURDER value was overwritten and then ignored in an else statement

The socket was closed, which didn't always trigger a return from
zmq_poll (even worse, the FD could theoretically get reused on a busy
server).  Switching to shutdown() guarantees operations will trigger an
error on that FD.

Per Justin's request, we don't return an error code on a single SIGTERM.
If you want it to die immediately with an error, either send two
SIGTERMs or just use SIGKILL (this is what SIGKILL is for anyways).