Make HAVEGE dependency optional
Recent mbedTLS versions will use getrandom(2) on linux, and people deploying in
a container may not need the chroot(), so urandom will be available.  In any
event many platform mbedTLS packages do not have HAVEGE, and it's only required
for SSL support, so we shouldn't hard depend on it.

This will now exit with an error message if the platform entropy
initialiazation fails and HAVEGE is not supported.

fixes #331