Fix crash when reading websocket packet after deliver task has stopped
Previously, conn->handler was set to NULL when the deliver task stopped. This was likely intended to prevent sending messages to the backend after the backend had sent a close message. However, the websocket code relies on conn->handler remaining non-NULL, and setting it to NULL only ever prevented sending leave notifications. So, let's just not set it to NULL, and use the deliver task status to decide whether to send leave notifications.